Move Machine to machine.js and command builder to command.js
